Change name and add icons
