+++ src/modules/resume/social-links/variants/basic/contributors.js
