+++ src/modules/resume/social-links/variants/basic/metadata.js
@@ -1,7 +1,7 @@
 
 const metadata = {
     name: "Basic",
-    description: "A very basic links variant",
+    description: "A very basic social links variant",
   }
   
   export default metadata;
